Add a better search functionality
In order to find a task it's full name had to be typed out.

This is inefficient for users especially if the task name is long.

Let's implement a find command that can list out tasks even if the keyword matches the task partially.

This leads to a more efficient search system.